Move targets out of framework (fuddly.framework.targets -> fuddly.targets)
