+++ src/fuddly/cli/templates/bare/targets.py
@@ -1,4 +1,4 @@
-from fuddly.framework.targets.debug import TestTarget
+from fuddly.targets.debug import TestTarget
 from fuddly.framework.target_helpers import EmptyTarget
 
 targets = [EmptyTarget(), TestTarget()]
