+++ src/fuddly/projects/generic/standard.py
@@ -25,9 +25,9 @@ import socket
 
 from fuddly.framework.comm_backends import Shell_Backend
 from fuddly.framework.plumbing import *
-from fuddly.framework.targets.local import LocalTarget
-from fuddly.framework.targets.network import NetworkTarget
-from fuddly.framework.targets.printer import PrinterTarget
+from fuddly.targets.local import LocalTarget
+from fuddly.targets.network import NetworkTarget
+from fuddly.targets.printer import PrinterTarget
 
 project = Project()
 project.default_dm = ['mydf','jpg']
