+++ src/fuddly/targets/network.py
